Renamed lists_database_manager (dart file) to database_manager
