+++ lib/view/list_widget.dart
@@ -1,5 +1,5 @@
 import 'package:collection/collection.dart';
-import 'package:lists/model/lists_database_manager.dart';
+import 'package:lists/model/database_manager.dart';
 import 'package:lists/model/list_model.dart';
 import 'package:flutter/material.dart';
 import 'package:lists/model/item.dart';
